+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 
 - feature: support reactive discovery client health indicator.
 - feature: Enhance default configuration to support `application*.yaml` and `bootstrap*.yaml`.
+- feat:adapt for nacos instance.